Add ability for Pi to able to manage bluetooth connections while running: - pair with new device in headless mode - create new bluetooth connection class bt_conn to own all functions related for managing the bluetooth connection - Add new dbus service that can be used to broadcast the status of bluetooth link to allow for asynchronous handling of bluetooth events
Added bluez simple-agent as a default paring agent in order to be able to pair in headless mode.

Change speed dial function so that the receive must be up. dialing_agent now distinguishes between single
number dialing and longer numbers

Change receiver event handler to read the real status of the receiver rather than toggle the status. This is peculiar to the
model of phone used as internal wiring meant that certain pabx functinos on the phone affect the receiver status.

Add function to refresh pulseaudio after a new bluetoth device has been added. This is a work around for a bug in pulseaudio
that mobile phones newly connected did not consistently appear as a card in pulseaudio.